requirements.txt
pydantic==1.8.2
python-dotenv==0.15.0
elasticsearch[async]~=7.13.1
uvicorn==0.13.0
fastapi==0.65.2
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
# opencv-contrib-python==4.5.5.64
PyYAML==5.4.1
# numpy==1.19.5
python-multipart==0.0.5
python-dateutil==2.8.2
pandas==1.1.3
python-socketio==4.6.0
websockets==9.1
python-engineio==3.13.2
Flask-SocketIO==4.3.1
openpyxl==3.0.9
python_arptable
xlsxwriter==3.0.3
requests==2.25.1
aiofiles==0.8.0
redis==4.3.4
editdistance
pyinstaller
auto-py-to-exe
